(big one) Fix race condition between Sonic Pi and jackd on boot
Firstly, the nc command was wrong for listening and outputting the
messages from the Sonic Pi log.

The bigger issue was that, after probably a day of debugging, the
addition of a long (15 second) sleep between jackd booting and the
jack_wait client connection eliminates the following error:

```
connect(2) call to /dev/shm/jack-4851/default/jack_0 failed
(err=Connection refused)
attempt to connect to server failed
```

It seems like accessing the jack server too quickly on boot causes a
crash - potentially the shm segment is corrupted somehow. This was very
difficult to ascertain as I thought it may be an environment specific issue with the Heroku
container runtime.